+++ app/src/main/java/com/piapps/flashcardpro/core/platform/component/bottom/BottomMenuItem.kt
@@ -2,4 +2,4 @@ package com.piapps.flashcardpro.core.platform.component.bottom
 
 import com.piapps.flashcardpro.R
 
-class BottomMenuItem(var id: Long, var text: String, var iconRes: Int = R.color.transparent)
+class BottomMenuItem(var id: Float, var text: String, var iconRes: Int = R.color.transparent)
